+++ changelog.md
@@ -7,3 +7,5 @@
 - Refrence Schema's via http and inline refrences
 - Support alternatives - oneOf, anyOf
 - Support for dependencies - array of conditionals
+- new Prop onError to get errors available in form on every change
+- new Prop uiData to control ui only data to add seperation of concern with form data